builds for d1_mini and native without flags

+++ lib/AsyncUnity/src/AsyncUnity/Depth.hpp
@@ -13,7 +13,7 @@ namespace AsyncUnity {
     
     public:
 
-      size_t highUsed = 0;
+      unsigned long highUsed = 0;
 
       const Error * push(const char * label, const long timeout) {
         if (MAX_DEPTH == _current) {
